18.05 class

+++ 3/4.4.py
@@ -4,10 +4,12 @@ b_russ = []
 
 with open('4.4_input.txt') as input_file:
     for line in input_file:
-        line = [float(i) for i in line.split(';')[1:]]
+        line = [int(i) for i in line.split(';')[1:]]
         b_math.append(line[0])
         b_phys.append(line[1])
         b_russ.append(line[2])
         print(sum(line) / len(line))
 
-print(sum(b_math) / len(b_math), sum(b_phys) / len(b_phys), sum(b_russ) / len(b_russ))
+print(sum(b_math) / len(b_math),
+      sum(b_phys) / len(b_phys),
+      sum(b_russ) / len(b_russ))
